# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Academic OS Integration API",
    "description": "Thin API Integration Facade connecting the React Frontend to the Academic OS Python Learning Intelligence Subsystem.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/students/{student_id}/profile": {
      "get": {
        "tags": [
          "Student Profile & Learning"
        ],
        "summary": "Get Profile",
        "description": "Retrieve the current student profile and concept mastery.",
        "operationId": "get_profile_students__student_id__profile_get",
        "parameters": [
          {
            "name": "student_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Student Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentProfileData"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/students/{student_id}/preferences": {
      "patch": {
        "tags": [
          "Student Profile & Learning"
        ],
        "summary": "Update Preferences",
        "description": "Update student learning preferences, course, or preferred language.",
        "operationId": "update_preferences_students__student_id__preferences_patch",
        "parameters": [
          {
            "name": "student_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Student Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StudentPreferencesUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentProfileData"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/students/{student_id}/learning": {
      "get": {
        "tags": [
          "Student Profile & Learning"
        ],
        "summary": "Get Learning State",
        "description": "Get aggregated learning state including Next Best Action,\npersonalized study plan, spaced repetition queue, and activity history.",
        "operationId": "get_learning_state_students__student_id__learning_get",
        "parameters": [
          {
            "name": "student_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Student Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LearningStateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/study-tools/generate": {
      "post": {
        "tags": [
          "Study Tools"
        ],
        "summary": "Generate Study Tools",
        "description": "Generate one or multiple personalized study tools from the 15 supported kinds:\n[explanation, summary, notes, study_guide, flashcards, quiz, exam, practice,\n code, coding_exercise, diagram, presentation, analogy, comparison, question_bank].\n- Grounded strictly in uploaded materials or trusted external evidence.\n- Zero model-memory fallback for unavailable evidence.\n- Supports SVG preview + PPTX download artifacts.",
        "operationId": "generate_study_tools_study_tools_generate_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "X-Student-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Student-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StudyToolsGenerateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudyToolsGenerateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/artifacts/{artifact_id}/download": {
      "get": {
        "tags": [
          "Artifacts"
        ],
        "summary": "Download Artifact",
        "description": "Download a generated artifact (e.g., SVG diagram, PPTX presentation).\nServes with proper MIME types and headers.",
        "operationId": "download_artifact_artifacts__artifact_id__download_get",
        "parameters": [
          {
            "name": "artifact_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Artifact Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "GeneratedResourceModel": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "content": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content"
          },
          "path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Path"
          },
          "filename": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Filename"
          },
          "download_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Download Url"
          },
          "mime_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mime Type"
          },
          "sources": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Sources"
          },
          "source_type": {
            "type": "string",
            "title": "Source Type",
            "default": "trusted_external"
          },
          "learner_state": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Learner State"
          },
          "teaching_strategy": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Teaching Strategy"
          },
          "evidence_limitation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evidence Limitation"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "type",
          "kind"
        ],
        "title": "GeneratedResourceModel"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LearningStateResponse": {
        "properties": {
          "profile": {
            "$ref": "#/components/schemas/StudentProfileData"
          },
          "next_action": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/NextBestActionData"
              },
              {
                "type": "null"
              }
            ]
          },
          "study_plan": {
            "items": {
              "$ref": "#/components/schemas/StudyPlanItem"
            },
            "type": "array",
            "title": "Study Plan"
          },
          "review_queue": {
            "items": {
              "$ref": "#/components/schemas/SpacedReviewItem"
            },
            "type": "array",
            "title": "Review Queue"
          },
          "learning_history": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Learning History"
          }
        },
        "type": "object",
        "required": [
          "profile"
        ],
        "title": "LearningStateResponse"
      },
...
      "NextBestActionData": {
        "properties": {
          "action": {
            "type": "string",
            "title": "Action"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          },
          "target_concepts": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Target Concepts"
          },
          "strategy": {
            "type": "string",
            "title": "Strategy"
          },
          "recommended_resources": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Recommended Resources"
          },
          "priority": {
            "type": "string",
            "title": "Priority"
          },
          "triggering_evidence": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Triggering Evidence"
          }
        },
        "type": "object",
        "required": [
          "action",
          "reason",
          "strategy",
          "priority"
        ],
        "title": "NextBestActionData"
      },
      "SpacedReviewItem": {
        "properties": {
          "concept": {
            "type": "string",
            "title": "Concept"
          },
          "due_date": {
            "type": "string",
            "title": "Due Date"
          },
          "interval_days": {
            "type": "number",
            "title": "Interval Days"
          },
          "repetition": {
            "type": "integer",
            "title": "Repetition"
          },
          "ease_factor": {
            "type": "number",
            "title": "Ease Factor"
          },
          "last_reviewed": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Reviewed"
          },
          "mastery": {
            "type": "number",
            "title": "Mastery",
            "default": 0.0
          }
        },
        "type": "object",
        "required": [
          "concept",
          "due_date",
          "interval_days",
          "repetition",
          "ease_factor"
        ],
        "title": "SpacedReviewItem"
      },
...
      "StudentPreferencesUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "course": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Course"
          },
          "preferred_language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preferred Language"
          },
          "learning_preference": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Learning Preference"
          }
        },
        "type": "object",
        "title": "StudentPreferencesUpdate"
      },
      "StudentProfileData": {
        "properties": {
          "student_id": {
            "type": "string",
            "title": "Student Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "course": {
            "type": "string",
            "title": "Course"
          },
          "preferred_language": {
            "type": "string",
            "title": "Preferred Language"
          },
          "learning_preference": {
            "type": "string",
            "title": "Learning Preference"
          },
          "concept_mastery": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Concept Mastery"
          },
          "weak_concepts": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Weak Concepts"
          },
          "unknown_concepts": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Unknown Concepts"
          },
          "strengths": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Strengths"
          },
          "misconceptions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Misconceptions"
          },
          "prerequisite_gaps": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Prerequisite Gaps"
          },
          "recent_topics": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Recent Topics"
          }
        },
        "type": "object",
        "required": [
          "student_id",
          "name",
          "course",
          "preferred_language",
          "learning_preference"
        ],
        "title": "StudentProfileData"
      },
      "StudyPlanItem": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "concept": {
            "type": "string",
            "title": "Concept"
          },
          "priority": {
            "type": "string",
            "title": "Priority"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          },
          "recommended_actions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Recommended Actions"
          },
          "completed": {
            "type": "boolean",
            "title": "Completed",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "title",
          "concept",
          "priority",
          "reason"
        ],
        "title": "StudyPlanItem"
      },
      "StudyToolsGenerateRequest": {
        "properties": {
          "student_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Student Id"
          },
          "topic": {
            "type": "string",
            "title": "Topic"
          },
          "kinds": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Kinds"
          },
          "language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language",
            "default": "en"
          },
          "document_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Document Ids"
          }
        },
        "type": "object",
        "required": [
          "topic",
          "kinds"
        ],
        "title": "StudyToolsGenerateRequest"
      },
      "StudyToolsGenerateResponse": {
        "properties": {
          "topic": {
            "type": "string",
            "title": "Topic"
          },
          "resources": {
            "additionalProperties": {
              "$ref": "#/components/schemas/GeneratedResourceModel"
            },
            "type": "object",
            "title": "Resources"
          }
        },
        "type": "object",
        "required": [
          "topic",
          "resources"
        ],
        "title": "StudyToolsGenerateResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}